Refactored for a more modular design.
This commit performs a refactor which dramatically improves on the
modularity of vim-surround. Now, by inheriting from BaseCommand,
Command classes can be instantiated which perform the functions
of a surround command in a very concise fashion.

This commit also uses disposables heavily in order to clean up in
a more idiomatic fashion. This preserves the responsible
configuration behavior.

This commit lays the groundwork for the deletion command. More
complexity may be required for the change surround command.